+++ cargo-dist/src/sign/macos.rs
@@ -16,9 +16,8 @@
 //! 2) Configure it to be usable for signing;
 //! 3) Import the certificate specified in the environment;
 //! 4) Actually perform the signing;
-//! 5) Let the keychain be deleted when the temporary directory is dropped.
-//!
-//! In the future, this module will also support notarization.
+//! 5) Submit the signed binary in a zip file to Apple for notarization;
+//! 6) Let the keychain be deleted when the temporary directory is dropped.
 use axoasset::LocalAsset;
 use axoprocess::Cmd;
 use base64::Engine;
@@ -131,10 +130,18 @@ struct CodesignEnv {
     pub identity: String,
     pub password: String,
     pub certificate: Vec<u8>,
+    pub notarization_apple_id: String,
+    pub notarization_password: String,
 }
 
 impl CodesignEnv {
-    pub fn from(identity: &str, password: &str, raw_certificate: &str) -> DistResult<Self> {
+    pub fn from(
+        identity: &str,
+        password: &str,
+        raw_certificate: &str,
+        notarization_apple_id: &str,
+        notarization_password: &str,
+    ) -> DistResult<Self> {
         let certificate = base64::prelude::BASE64_STANDARD
             .decode(raw_certificate)
             .map_err(|_| DistError::CertificateDecodeError {})?;
@@ -143,6 +150,8 @@ impl CodesignEnv {
             identity: identity.to_owned(),
             password: password.to_owned(),
             certificate,
+            notarization_apple_id: notarization_apple_id.to_owned(),
+            notarization_password: notarization_password.to_owned(),
         })
     }
 }
@@ -153,6 +162,8 @@ impl std::fmt::Debug for CodesignEnv {
             .field("identity", &"<hidden>")
             .field("password", &"<hidden>")
             .field("certificate", &"<hidden>")
+            .field("notarization_apple_id", &"<hidden>")
+            .field("notarization_password", &"<hidden>")
             .finish()
     }
 }
@@ -163,12 +174,26 @@ impl Codesign {
             return Ok(None);
         }
 
-        if let (Some(identity), Some(password), Some(certificate)) = (
+        if let (
+            Some(identity),
+            Some(password),
+            Some(certificate),
+            Some(notarization_apple_id),
+            Some(notarization_password),
+        ) = (
             Self::var("CODESIGN_IDENTITY"),
             Self::var("CODESIGN_CERTIFICATE_PASSWORD"),
             Self::var("CODESIGN_CERTIFICATE"),
+            Self::var("CODESIGN_NOTARIZATION_APPLE_ID"),
+            Self::var("CODESIGN_NOTARIZATION_PASSWORD"),
         ) {
-            let env = CodesignEnv::from(&identity, &password, &certificate)?;
+            let env = CodesignEnv::from(
+                &identity,
+                &password,
+                &certificate,
+                &notarization_apple_id,
+                &notarization_password,
+            )?;
 
             Ok(Some(Self { env }))
         } else {
@@ -184,18 +209,62 @@ impl Codesign {
         val
     }
 
-    pub fn sign(&self, file: &Utf8Path) -> DistResult<()> {
-        let password = uuid::Uuid::new_v4().as_hyphenated().to_string();
-        let keychain = Keychain::create(password)?;
-        keychain.import_certificate(&self.env.certificate, &self.env.password)?;
-
+    fn sign_binary(&self, file: &Utf8Path, keychain: &Keychain) -> DistResult<()> {
         let mut cmd = Cmd::new("/usr/bin/codesign", "sign macOS artifacts");
         cmd.arg("--sign").arg(&self.env.identity);
         cmd.arg("--keychain").arg(&keychain.path);
+        cmd.arg("--timestamp");
+        cmd.arg("--options").arg("runtime");
         cmd.arg(file);
         cmd.stdout_to_stderr();
         cmd.output()?;
 
         Ok(())
+    }
+
+    fn notarize(&self, file: &Utf8Path) -> DistResult<()> {
+        let (_root, root_path) = create_tmp()?;
+        let mut file_name: Utf8PathBuf = file
+            .file_name()
+            .expect("path to notarize had no file name")
+            .into();
+        file_name.set_extension("zip");
+        let zip_path = root_path.join(file_name);
+
+        // A bare Mach-O binary cannot be notarized. A package or zip must be sent.
+        let mut ditto_cmd = Cmd::new("/usr/bin/ditto", "zip macOS artifacts for notarization");
+        ditto_cmd.arg("-c");
+        ditto_cmd.arg("-k");
+        ditto_cmd.arg("--keepParent");
+        ditto_cmd.arg(file);
+        ditto_cmd.arg(&zip_path);
+        ditto_cmd.stdout_to_stderr();
+        ditto_cmd.output()?;
+
+        // xcrun notarytool submit --apple-id will@ox1de.com --team-id 4459W66H65 --password rdus-eanr-deij-wbnc --wait dist.zip
+        let mut notarize_cmd = Cmd::new("/usr/bin/xcrun", "notarize macOS artifacts");
+        notarize_cmd.arg("notarytool");
+        notarize_cmd.arg("submit");
+        notarize_cmd
+            .arg("--apple-id")
+            .arg(&self.env.notarization_apple_id);
+        notarize_cmd.arg("--team-id").arg(&self.env.identity);
+        notarize_cmd
+            .arg("--password")
+            .arg(&self.env.notarization_password);
+        notarize_cmd.arg(&zip_path);
+        notarize_cmd.stdout_to_stderr();
+        notarize_cmd.output()?;
+
+        Ok(())
+    }
+
+    pub fn sign(&self, file: &Utf8Path) -> DistResult<()> {
+        let password = uuid::Uuid::new_v4().as_hyphenated().to_string();
+        let keychain = Keychain::create(password)?;
+        keychain.import_certificate(&self.env.certificate, &self.env.password)?;
+
+        self.sign_binary(file, &keychain)?;
+        self.notarize(file)
     }
 }