+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Nothing Yet!
 
 # Version 1.0.3 (2025-07-02)
 
-- Enable macOS binary signing for `dist` itself.
+- Fix macOS binary signing not triggering despite being enabled.
+- Submit codesigned macOS binaries to Apple for notarization.
 
 # Version 1.0.2 (2025-06-19)